Split into shell specific collections
