+++ tests/shell/__init__.py
@@ -1,0 +1,262 @@
+# Copyright (C) 2012 Anaconda, Inc
+# SPDX-License-Identifier: BSD-3-Clause
+from __future__ import annotations
+
+import os
+import sys
+from os.path import dirname
+from shutil import which
+from signal import SIGINT
+from typing import TYPE_CHECKING
+from uuid import uuid4
+
+from pexpect.popen_spawn import PopenSpawn
+
+from conda import CONDA_PACKAGE_ROOT, CONDA_SOURCE_ROOT
+from conda.activate import activator_map, native_path_to_unix
+from conda.common.compat import on_win
+from conda.utils import quote_for_shell
+
+if TYPE_CHECKING:
+    from collections.abc import Iterable
+
+
+# Here, by removing --dev you can try weird situations that you may want to test, upgrade paths
+# and the like? What will happen is that the conda being run and the shell scripts it's being run
+# against will be essentially random and will vary over the course of activating and deactivating
+# environments. You will have absolutely no idea what's going on as you change code or scripts and
+# encounter some different code that ends up being run (some of the time). You will go slowly mad.
+# No, you are best off keeping --dev on the end of these. For sure, if conda bundled its own tests
+# module then we could remove --dev if we detect we are being run in that way.
+dev_arg = "--dev"
+activate = f" activate {dev_arg} "
+deactivate = f" deactivate {dev_arg} "
+install = f" install {dev_arg} "
+
+# hdf5 version to use in tests
+HDF5_VERSION = "1.12.1"
+
+
+class InteractiveShellType(type):
+    EXE = quote_for_shell(native_path_to_unix(sys.executable))
+    SHELLS: dict[str, dict] = {
+        "posix": {
+            "activator": "posix",
+            "init_command": (
+                f'eval "$({EXE} -m conda shell.posix hook {dev_arg})" '
+                # want CONDA_SHLVL=0 before running tests so deactivate any active environments
+                # since we do not know how many environments have been activated by the user/CI
+                # just to be safe deactivate a few times
+                "&& conda deactivate "
+                "&& conda deactivate "
+                "&& conda deactivate "
+                "&& conda deactivate"
+            ),
+            "print_env_var": 'echo "$%s"',
+        },
+        "bash": {
+            # MSYS2's login scripts handle mounting the filesystem. Without it, /c is /cygdrive.
+            "args": ("-l",) if on_win else (),
+            "base_shell": "posix",  # inheritance implemented in __init__
+        },
+        "dash": {"base_shell": "posix"},
+        "zsh": {"base_shell": "posix"},
+        # It should be noted here that we use the latest hook with whatever conda.exe is installed
+        # in sys.prefix (and we will activate all of those PATH entries).  We will set PYTHONPATH
+        # though, so there is that.  What I'm getting at is that this is a huge mixup and a mess.
+        "cmd.exe": {
+            "activator": "cmd.exe",
+            # For non-dev-mode you'd have:
+            #            'init_command': 'set "CONDA_SHLVL=" '
+            #                            '&& @CALL {}\\shell\\condabin\\conda_hook.bat {} '
+            #                            '&& set CONDA_EXE={}'
+            #                            '&& set _CE_M='
+            #                            '&& set _CE_CONDA='
+            #                            .format(CONDA_PACKAGE_ROOT, dev_arg,
+            #                                    join(sys.prefix, "Scripts", "conda.exe")),
+            "init_command": (
+                '@SET "CONDA_SHLVL=" '
+                f"&& @CALL {CONDA_PACKAGE_ROOT}\\shell\\condabin\\conda_hook.bat {dev_arg} "
+                f'&& @SET "CONDA_EXE={sys.executable}" '
+                '&& @SET "_CE_M=-m" '
+                '&& @SET "_CE_CONDA=conda"'
+            ),
+            "print_env_var": "@ECHO %%%s%%",
+        },
+        "csh": {
+            "activator": "csh",
+            # Trying to use -x with `tcsh` on `macOS` results in some problems:
+            # This error from `PyCharm`:
+            # BrokenPipeError: [Errno 32] Broken pipe (writing to self.proc.stdin).
+            # .. and this one from the `macOS` terminal:
+            # pexpect.exceptions.EOF: End Of File (EOF).
+            # 'args': ('-x',),
+            "init_command": (
+                f'set _CONDA_EXE="{CONDA_PACKAGE_ROOT}/shell/bin/conda"; '
+                f"source {CONDA_PACKAGE_ROOT}/shell/etc/profile.d/conda.csh;"
+            ),
+            "print_env_var": 'echo "$%s"',
+        },
+        "tcsh": {"base_shell": "csh"},
+        "fish": {
+            "activator": "fish",
+            "init_command": f"eval ({EXE} -m conda shell.fish hook {dev_arg})",
+            "print_env_var": "echo $%s",
+        },
+        # We don't know if the PowerShell executable is called
+        # powershell, pwsh, or pwsh-preview.
+        "powershell": {
+            "activator": "powershell",
+            "args": ("-NoProfile", "-NoLogo"),
+            "init_command": (
+                f"{sys.executable} -m conda shell.powershell hook --dev "
+                "| Out-String "
+                "| Invoke-Expression "
+                # want CONDA_SHLVL=0 before running tests so deactivate any active environments
+                # since we do not know how many environments have been activated by the user/CI
+                # just to be safe deactivate a few times
+                "; conda deactivate "
+                "; conda deactivate "
+                "; conda deactivate "
+                "; conda deactivate"
+            ),
+            "print_env_var": "$Env:%s",
+            "exit_cmd": "exit",
+        },
+        "pwsh": {"base_shell": "powershell"},
+        "pwsh-preview": {"base_shell": "powershell"},
+    }
+
+    def __call__(self, shell_name: str, **kwargs):
+        return super().__call__(
+            shell_name,
+            **{
+                **self.SHELLS.get(self.SHELLS[shell_name].get("base_shell"), {}),
+                **self.SHELLS[shell_name],
+                **kwargs,
+            },
+        )
+
+
+class InteractiveShell(metaclass=InteractiveShellType):
+    def __init__(
+        self,
+        shell_name: str,
+        *,
+        activator: str,
+        args: Iterable[str] = (),
+        init_command: str,
+        print_env_var: str,
+        exit_cmd: str | None = None,
+        base_shell: str | None = None,  # ignored
+        shell_path: str | None = None,
+    ):
+        self.shell_name = shell_name
+        if not shell_path:
+            assert (shell_path := which(shell_name))
+        self.shell_exe = quote_for_shell(shell_path, *args)
+        self.shell_dir = dirname(shell_path)
+
+        self.activator = activator_map[activator]()
+        self.args = args
+        self.init_command = init_command
+        self.print_env_var = print_env_var
+        self.exit_cmd = exit_cmd
+
+    def __enter__(self):
+        self.p = PopenSpawn(
+            self.shell_exe,
+            timeout=30,
+            maxread=5000,
+            searchwindowsize=None,
+            logfile=sys.stdout,
+            cwd=os.getcwd(),
+            env={
+                **os.environ,
+                "CONDA_AUTO_ACTIVATE_BASE": "false",
+                "CONDA_AUTO_STACK": "0",
+                "CONDA_CHANGEPS1": "true",
+                # "CONDA_ENV_PROMPT": "({default_env}) ",
+                "PYTHONPATH": self.path_conversion(CONDA_SOURCE_ROOT),
+                "PATH": self.activator.pathsep_join(
+                    self.path_conversion(
+                        (
+                            *self.activator._get_starting_path_list(),
+                            self.shell_dir,
+                        )
+                    )
+                ),
+                # ensure PATH is shared with any msys2 bash shell, rather than starting fresh
+                "MSYS2_PATH_TYPE": "inherit",
+                "CHERE_INVOKING": "1",
+            },
+            encoding="utf-8",
+            codec_errors="strict",
+        )
+
+        if self.init_command:
+            self.p.sendline(self.init_command)
+
+        self.clear()
+
+        return self
+
+    def __exit__(self, exc_type, exc_val, exc_tb):
+        if exc_type is not None:
+            print(f"Exception encountered: ({exc_type}) {exc_val}", file=sys.stderr)
+
+        if self.p:
+            if self.exit_cmd:
+                self.sendline(self.exit_cmd)
+
+            self.p.kill(SIGINT)
+
+    def sendline(self, *args, **kwargs):
+        return self.p.sendline(*args, **kwargs)
+
+    def expect(self, *args, **kwargs):
+        try:
+            return self.p.expect(*args, **kwargs)
+        except Exception:
+            print(f"{self.p.before=}", file=sys.stderr)
+            print(f"{self.p.after=}", file=sys.stderr)
+            raise
+
+    def expect_exact(self, *args, **kwargs):
+        try:
+            return self.p.expect_exact(*args, **kwargs)
+        except Exception:
+            print(f"{self.p.before=}", file=sys.stderr)
+            print(f"{self.p.after=}", file=sys.stderr)
+            raise
+
+    def assert_env_var(self, env_var, value, use_exact=False):
+        # value is actually a regex
+        self.sendline(self.print_env_var % env_var)
+        if use_exact:
+            self.expect_exact(value)
+            self.clear()
+        else:
+            self.expect(rf"{value}\r?\n")
+
+    def get_env_var(self, env_var, default=None):
+        self.sendline(self.print_env_var % env_var)
+        if self.shell_name == "cmd.exe":
+            self.expect(rf"@ECHO %{env_var}%\r?\n([^\r\n]*)\r?\n")
+        elif self.shell_name in ("powershell", "pwsh"):
+            self.expect(rf"\$Env:{env_var}\r?\n([^\r\n]*)\r?\n")
+        else:
+            marker = f"get_env_var-{uuid4().hex}"
+            self.sendline(f"echo {marker}")
+            self.expect(rf"([^\r\n]*)\r?\n{marker}\r?\n")
+
+        value = self.p.match.group(1)
+        return default if value is None else value
+
+    def clear(self) -> None:
+        marker = f"clear-{uuid4().hex}"
+        self.sendline(f"echo {marker}")
+        self.expect(rf"{marker}\r?\n")
+
+    def path_conversion(self, *args, **kwargs):
+        return self.activator.path_conversion(*args, **kwargs)
